requirements.txt
geocoder
django-map-widgets==0.2.2
telethon==1.7.7
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.2
Markdown==3.0.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
https://github.com/darklow/django-suit/tarball/v2
